miac/requirements.txt
Django==3.2.4
djangorestframework==3.12.4
dj-rest-auth==2.1.7
django-allauth==0.44.0
djangorestframework-simplejwt==4.7.1
django-environ==0.4.5
django-redis==5.0.0
psycopg2-binary==2.8.6
django-phonenumber-field==5.2.0
drf-yasg==1.20.0
gunicorn==20.1.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
factory-boy==3.2.0
